Allow modified `LegacyDefaultSkin`s to fallback to the non-modified version
We have fixed the ability for skins to fallback to resources from legacy
skins a while back, but some limitations were still left in place. This
fixes two such limitations:

- The components provided by the legacy skin were only returning classic
  components when a certain font was present. This is no longer required
and as such has been removed.
- Until now, modifying the default skin in the editor would always use
  the legacy resources provided by `osu-resources`, which meant that new
assets could not be sourced from the modified skin (as it is defined in
the database as a `LegacyDefaultSkin`). Now all instantiated
`LegacyDefaultSkins` use a local resource store, relying on the fallback
provided by `SkinManager` to access the fallback resources.