Move all incoming multiplayer signalr events to a single task chain
Hopefully a relatively temporary fix for this issue, but our goal is
first and foremost to make this reliable. Due to the gap found during
the join / quit sequence where the `Room` may not be set (as it is set
on the update thread *after* an async fetch operation) this is the only
reliable way to make thins work.

A future solution will likely see moving all signalr interactions to the
*signalr* thread instead, and exposing `Room` using a thread-safe
method.