Refactor `BeatmapDownloadTracker` to handle cases of local outdated beatmaps
This was some work I started with the original intention of adding a new
`DownloadState` for server-side modifications available. I ended up not
requiring it, but I think there's still benefit in getting this PR'd as
(from my recollection) it fixes some edge case bugs and makes the code
easier to visually parse.