Allow distance snap precision adjust using shift key
By default, using ctrl+alt to adjust distance snap will now adjust by 0.01 instead of 0.1. Holding shift as well will allow adjustments of 0.1 as before.

This is done to match other shift-adjustments which always *increase* the amount of adjustment, rather than decrease. It is opposite of stable, which does it back-to-front for whatever reason.

- [ ] Depends on framework fix for shift-mousewheel flipping