Fix release time auto generator
But really, to make this work the concept of "key up" events likely
needs to be split out, so it doesn't hold up movement to the next
hitobject while in the keydown state.

Not too sure how to fix this without a completely rewrite of auto
generation. Or maybe even reapproaching the whole thing (do we want to
use static replays for auto generation? this won't ever give the
behaviour people have come to expect in autopilot, at very least).