Remove `GlobalActionContainer`'s prioritised status
While this makes sense on a surface level, it leads to some very
restrictive behaviour which keeps blocking forward progress.

Namely, with this flag in place, it will handle input before *raw*
key/mouse events, which goes against expectations. It means that
bindings to global actions on mouse buttons will override expected local
behaviours in weird ways.

The straw-that-broke-the-camel's-back here was attempting to make
absolute scroll at song select rebindable. This broke anything which
handles right click actions, because the global container now has
prioritised handling of right mouse.

I believe we should not use this flag for such a general enum as
`GlobalAction`. It makes sense for `PlatformAction` and may make sense
for some of the contents of `GlobalAction`, but as touched on
[previously](https://github.com/ppy/osu/commit/5454d1caa1927428fe339e3d5b47ddfe45f9dfb6#diff-e0a108aa86713360cbe2541a78dc0dd100d2a5e91753f42139017b2b3842db0dL172),
as long as we are using this enum for actions which are not always
global, it doesn't make sense.