Add context menus to new carousel panels
This changes the flow of actions (yet again.. sorry) to standardise what
is show on the beatmap options button in the footer and the context
menus.

This is something I wanted from the start – for devices where right
click is not available, it's always preferrable to have a second method
of accessing actions which isn't the context menu.